// src/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & {
  [P in K]-?: NonNullable<T[P]>;
};
// ...
export type HeatMap = {
  __typename?: 'HeatMap';
  data: Array<Scalars['Int']>;
  end: Scalars['Date'];
  start: Scalars['Date'];
};

export type Mutation = {
  __typename?: 'Mutation';
  /** Add a target */
  targetCreate: Maybe<Target>;
  /** Delete a target */
  targetDelete: Maybe<Scalars['Boolean']>;
  /** Update a target */
  targetUpdate: Maybe<Target>;
  /** Add timePiece for particular target */
  timePieceCreate: Maybe<TimePiece>;
  /** Delete a timePiece */
  timePieceDelete: Maybe<Scalars['Boolean']>;
  /** Update a timePiece */
  timePieceUpdate: Maybe<TimePiece>;
  /** Add timePieces for particular target */
  timePiecesCreateForTarget: Maybe<Array<TimePiece>>;
  /** Add a user */
  userCreate: Maybe<User>;
  /** Update a user */
  userUpdate: Maybe<User>;
};

export type MutationTargetCreateArgs = {
  input: TargetCreateInput;
  user_id: Scalars['ID'];
};

export type MutationTargetDeleteArgs = {
  target_id: Scalars['ID'];
};

export type MutationTargetUpdateArgs = {
  input: TargetUpdateInput;
  target_id: Scalars['ID'];
};

export type MutationTimePieceCreateArgs = {
  input: TimePieceCreateInput;
  target_id: Scalars['String'];
};

export type MutationTimePieceDeleteArgs = {
  timepiece_id: Scalars['Int'];
};

export type MutationTimePieceUpdateArgs = {
  input: TimePieceUpdateInput;
  timepiece_id: Scalars['Int'];
};

export type MutationTimePiecesCreateForTargetArgs = {
  input: Array<TimePieceCreateInput>;
  target_id: Scalars['String'];
};

export type MutationUserCreateArgs = {
  input: UserCreateInput;
};

export type MutationUserUpdateArgs = {
  input: UserUpdateInput;
  user_id: Scalars['ID'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Test Connection */
  test: Scalars['Boolean'];
  /** Try to find one user */
  user: User;
};

export type QueryUserArgs = {
  username: Scalars['String'];
};

/** One Target */
export type Target = {
  __typename?: 'Target';
  /** Create Time */
  created_at: Scalars['DateTime'];
  /** Generate HeatMap */
  heatMap: HeatMap;
  /** Target ID */
  id: Scalars['ID'];
  /** Timepieces in the last week */
  lastWeekTimePieces: Array<TimePiece>;
  /** Target name */
  name: Scalars['String'];
  /** Total pomodoro count */
  pomodoroCount: Scalars['Int'];
  /** Consumption time periods */
  timePieces: TimePieceConnection;
  /** Time spent on this target */
  timeSpent: Scalars['Int'];
  /** Today pomodoro count */
  todayPomodoroCount: Scalars['Int'];
  /** Today focus time length */
  todayTimeSpent: Scalars['Int'];
  /** Target type */
  type: Maybe<TargetType>;
};
// ...
export enum TargetType {
  Longterm = 'LONGTERM',
  Normal = 'NORMAL',
}

export type TargetUpdateInput = {
  name: InputMaybe<Scalars['String']>;
  timeSpent: InputMaybe<Scalars['Int']>;
  type: InputMaybe<TargetType>;
};

/** One Time Piece */
export type TimePiece = {
  __typename?: 'TimePiece';
  /** Time spend on this piece, count as seconds */
  duration: Scalars['Int'];
  /** TimePiece ID */
  id: Scalars['Int'];
  /** Piece start time */
  start: Scalars['DateTime'];
  /** Target this piece belongs to */
  target: Target;
  /** Timepiece type */
  type: Maybe<TimePieceType>;
};

export type TimePieceConnection = {
  __typename?: 'TimePieceConnection';
  edges: Array<TimePieceEdge>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int'];
};

export type TimePieceCreateInput = {
  duration: Scalars['Int'];
  start: Scalars['DateTime'];
  type: InputMaybe<TimePieceType>;
};

export type TimePieceEdge = {
  __typename?: 'TimePieceEdge';
  cursor: Scalars['String'];
  node: TimePiece;
};

export enum TimePieceType {
  Normal = 'NORMAL',
  Pomodoro = 'POMODORO',
}

export type TimePieceUpdateInput = {
  duration: InputMaybe<Scalars['Int']>;
  start: InputMaybe<Scalars['DateTime']>;
  type: InputMaybe<TimePieceType>;
};

/** User */
export type User = {
  __typename?: 'User';
  /** Create Time */
  created_at: Scalars['DateTime'];
  /** Generate heatMap */
  heatMap: HeatMap;
  /** User ID */
  id: Scalars['ID'];
  /** Timepieces in the last week */
  lastWeekTimePieces: Array<TimePiece>;
  /** Total pomodoro count */
  pomodoroCount: Scalars['Int'];
  /** Find a target by ID */
  target: Target;
  /** Targets */
  targets: Maybe<Array<Target>>;
  /** Consumption time periods */
  timePieces: TimePieceConnection;
  /** Total focus time length */
  timeSpent: Scalars['Int'];
  /** Today pomodoro count */
  todayPomodoroCount: Scalars['Int'];
  /** Today focus time length */
  todayTimeSpent: Scalars['Int'];
  /** Username */
  username: Scalars['String'];
};
// ...
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;

export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type ResolversTypes = ResolversObject<{
  BigInt: ResolverTypeWrapper<Scalars['BigInt']>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  Byte: ResolverTypeWrapper<Scalars['Byte']>;
  Currency: ResolverTypeWrapper<Scalars['Currency']>;
  Date: ResolverTypeWrapper<Scalars['Date']>;
  DateTime: ResolverTypeWrapper<Scalars['DateTime']>;
  EmailAddress: ResolverTypeWrapper<Scalars['EmailAddress']>;
  GUID: ResolverTypeWrapper<Scalars['GUID']>;
  HSL: ResolverTypeWrapper<Scalars['HSL']>;
  HSLA: ResolverTypeWrapper<Scalars['HSLA']>;
  HeatMap: ResolverTypeWrapper<HeatMap>;
  HexColorCode: ResolverTypeWrapper<Scalars['HexColorCode']>;
  Hexadecimal: ResolverTypeWrapper<Scalars['Hexadecimal']>;
  IBAN: ResolverTypeWrapper<Scalars['IBAN']>;
  ID: ResolverTypeWrapper<Scalars['ID']>;
  IPv4: ResolverTypeWrapper<Scalars['IPv4']>;
  IPv6: ResolverTypeWrapper<Scalars['IPv6']>;
  ISBN: ResolverTypeWrapper<Scalars['ISBN']>;
  ISO8601Duration: ResolverTypeWrapper<Scalars['ISO8601Duration']>;
  Int: ResolverTypeWrapper<Scalars['Int']>;
  JSON: ResolverTypeWrapper<Scalars['JSON']>;
  JSONObject: ResolverTypeWrapper<Scalars['JSONObject']>;
  JWT: ResolverTypeWrapper<Scalars['JWT']>;
  Latitude: ResolverTypeWrapper<Scalars['Latitude']>;
  LocalDate: ResolverTypeWrapper<Scalars['LocalDate']>;
  LocalEndTime: ResolverTypeWrapper<Scalars['LocalEndTime']>;
  LocalTime: ResolverTypeWrapper<Scalars['LocalTime']>;
  Long: ResolverTypeWrapper<Scalars['Long']>;
  Longitude: ResolverTypeWrapper<Scalars['Longitude']>;
  MAC: ResolverTypeWrapper<Scalars['MAC']>;
  Mutation: ResolverTypeWrapper<{}>;
  NegativeFloat: ResolverTypeWrapper<Scalars['NegativeFloat']>;
  NegativeInt: ResolverTypeWrapper<Scalars['NegativeInt']>;
  NonEmptyString: ResolverTypeWrapper<Scalars['NonEmptyString']>;
  NonNegativeFloat: ResolverTypeWrapper<Scalars['NonNegativeFloat']>;
  NonNegativeInt: ResolverTypeWrapper<Scalars['NonNegativeInt']>;
  NonPositiveFloat: ResolverTypeWrapper<Scalars['NonPositiveFloat']>;
  NonPositiveInt: ResolverTypeWrapper<Scalars['NonPositiveInt']>;
  ObjectID: ResolverTypeWrapper<Scalars['ObjectID']>;
  OrderDirection: OrderDirection;
  PageInfo: ResolverTypeWrapper<PageInfo>;
  PhoneNumber: ResolverTypeWrapper<Scalars['PhoneNumber']>;
  Port: ResolverTypeWrapper<Scalars['Port']>;
  PositiveFloat: ResolverTypeWrapper<Scalars['PositiveFloat']>;
  PositiveInt: ResolverTypeWrapper<Scalars['PositiveInt']>;
  PostalCode: ResolverTypeWrapper<Scalars['PostalCode']>;
  Query: ResolverTypeWrapper<{}>;
  RGB: ResolverTypeWrapper<Scalars['RGB']>;
  RGBA: ResolverTypeWrapper<Scalars['RGBA']>;
  SafeInt: ResolverTypeWrapper<Scalars['SafeInt']>;
  String: ResolverTypeWrapper<Scalars['String']>;
  Target: ResolverTypeWrapper<Target>;
  TargetCreateInput: TargetCreateInput;
  TargetType: TargetType;
  TargetUpdateInput: TargetUpdateInput;
  Time: ResolverTypeWrapper<Scalars['Time']>;
  TimePiece: ResolverTypeWrapper<TimePiece>;
  TimePieceConnection: ResolverTypeWrapper<TimePieceConnection>;
  TimePieceCreateInput: TimePieceCreateInput;
  TimePieceEdge: ResolverTypeWrapper<TimePieceEdge>;
  TimePieceType: TimePieceType;
  TimePieceUpdateInput: TimePieceUpdateInput;
  Timestamp: ResolverTypeWrapper<Scalars['Timestamp']>;
  URL: ResolverTypeWrapper<Scalars['URL']>;
  USCurrency: ResolverTypeWrapper<Scalars['USCurrency']>;
  UUID: ResolverTypeWrapper<Scalars['UUID']>;
  UnsignedFloat: ResolverTypeWrapper<Scalars['UnsignedFloat']>;
  UnsignedInt: ResolverTypeWrapper<Scalars['UnsignedInt']>;
  User: ResolverTypeWrapper<User>;
  UserCreateInput: UserCreateInput;
  UserUpdateInput: UserUpdateInput;
  UtcOffset: ResolverTypeWrapper<Scalars['UtcOffset']>;
  Void: ResolverTypeWrapper<Scalars['Void']>;
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
  BigInt: Scalars['BigInt'];
  Boolean: Scalars['Boolean'];
  Byte: Scalars['Byte'];
  Currency: Scalars['Currency'];
  Date: Scalars['Date'];
  DateTime: Scalars['DateTime'];
  EmailAddress: Scalars['EmailAddress'];
  GUID: Scalars['GUID'];
  HSL: Scalars['HSL'];
  HSLA: Scalars['HSLA'];
  HeatMap: HeatMap;
  HexColorCode: Scalars['HexColorCode'];
  Hexadecimal: Scalars['Hexadecimal'];
  IBAN: Scalars['IBAN'];
  ID: Scalars['ID'];
  IPv4: Scalars['IPv4'];
  IPv6: Scalars['IPv6'];
  ISBN: Scalars['ISBN'];
  ISO8601Duration: Scalars['ISO8601Duration'];
  Int: Scalars['Int'];
  JSON: Scalars['JSON'];
  JSONObject: Scalars['JSONObject'];
  JWT: Scalars['JWT'];
  Latitude: Scalars['Latitude'];
  LocalDate: Scalars['LocalDate'];
  LocalEndTime: Scalars['LocalEndTime'];
  LocalTime: Scalars['LocalTime'];
  Long: Scalars['Long'];
  Longitude: Scalars['Longitude'];
  MAC: Scalars['MAC'];
  Mutation: {};
  NegativeFloat: Scalars['NegativeFloat'];
  NegativeInt: Scalars['NegativeInt'];
  NonEmptyString: Scalars['NonEmptyString'];
  NonNegativeFloat: Scalars['NonNegativeFloat'];
  NonNegativeInt: Scalars['NonNegativeInt'];
  NonPositiveFloat: Scalars['NonPositiveFloat'];
  NonPositiveInt: Scalars['NonPositiveInt'];
  ObjectID: Scalars['ObjectID'];
  PageInfo: PageInfo;
  PhoneNumber: Scalars['PhoneNumber'];
  Port: Scalars['Port'];
  PositiveFloat: Scalars['PositiveFloat'];
  PositiveInt: Scalars['PositiveInt'];
  PostalCode: Scalars['PostalCode'];
  Query: {};
  RGB: Scalars['RGB'];
  RGBA: Scalars['RGBA'];
  SafeInt: Scalars['SafeInt'];
  String: Scalars['String'];
  Target: Target;
  TargetCreateInput: TargetCreateInput;
  TargetUpdateInput: TargetUpdateInput;
  Time: Scalars['Time'];
  TimePiece: TimePiece;
  TimePieceConnection: TimePieceConnection;
  TimePieceCreateInput: TimePieceCreateInput;
  TimePieceEdge: TimePieceEdge;
  TimePieceUpdateInput: TimePieceUpdateInput;
  Timestamp: Scalars['Timestamp'];
  URL: Scalars['URL'];
  USCurrency: Scalars['USCurrency'];
  UUID: Scalars['UUID'];
  UnsignedFloat: Scalars['UnsignedFloat'];
  UnsignedInt: Scalars['UnsignedInt'];
  User: User;
  UserCreateInput: UserCreateInput;
  UserUpdateInput: UserUpdateInput;
  UtcOffset: Scalars['UtcOffset'];
  Void: Scalars['Void'];
}>;
// ...
export type MutationResolvers<
  ContextType = any,
  ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']
> = ResolversObject<{
  targetCreate: Resolver<
    Maybe<ResolversTypes['Target']>,
    ParentType,
    ContextType,
    RequireFields<MutationTargetCreateArgs, 'input' | 'user_id'>
  >;
  targetDelete: Resolver<
    Maybe<ResolversTypes['Boolean']>,
    ParentType,
    ContextType,
    RequireFields<MutationTargetDeleteArgs, 'target_id'>
  >;
  targetUpdate: Resolver<
    Maybe<ResolversTypes['Target']>,
    ParentType,
    ContextType,
    RequireFields<MutationTargetUpdateArgs, 'input' | 'target_id'>
  >;
  timePieceCreate: Resolver<
    Maybe<ResolversTypes['TimePiece']>,
    ParentType,
    ContextType,
    RequireFields<MutationTimePieceCreateArgs, 'input' | 'target_id'>
  >;
  timePieceDelete: Resolver<
    Maybe<ResolversTypes['Boolean']>,
    ParentType,
    ContextType,
    RequireFields<MutationTimePieceDeleteArgs, 'timepiece_id'>
  >;
  timePieceUpdate: Resolver<
    Maybe<ResolversTypes['TimePiece']>,
    ParentType,
    ContextType,
    RequireFields<MutationTimePieceUpdateArgs, 'input' | 'timepiece_id'>
  >;
  timePiecesCreateForTarget: Resolver<
    Maybe<Array<ResolversTypes['TimePiece']>>,
    ParentType,
    ContextType,
    RequireFields<MutationTimePiecesCreateForTargetArgs, 'input' | 'target_id'>
  >;
  userCreate: Resolver<
    Maybe<ResolversTypes['User']>,
    ParentType,
    ContextType,
    RequireFields<MutationUserCreateArgs, 'input'>
  >;
  userUpdate: Resolver<
    Maybe<ResolversTypes['User']>,
    ParentType,
    ContextType,
    RequireFields<MutationUserUpdateArgs, 'input' | 'user_id'>
  >;
}>;